refactor(handler): use getIcon() from handler
